+++ web/src/components/survey/question-types/custom-date-type.ts
@@ -1,0 +1,41 @@
+import Vue from "vue";
+import CustomDate from "../components/CustomDate.vue";
+import { WidgetValueName } from "./question-types";
+
+export function initCustomDate(Survey: any) {
+  const widget = {
+    name: WidgetValueName[WidgetValueName.CustomDate],
+    title: "Date Input",
+    iconName: "icon-date",
+    widgetIsLoaded: function() {
+      return true;
+    },
+    isFit: function(question: any) {
+      return question.getType() === "customdate";
+    },
+    htmlTemplate: "<div></div>",
+    activatedByChanged: function(activatedBy: any) {
+      Survey.JsonObject.metaData.addClass("customdate", [], null, "empty");
+      Survey.JsonObject.metaData.addProperties("customdate", [
+        {
+          name: "dateYearsAhead:number",
+          default: 0
+        },
+        {
+          name: "dateYearsBehind:number",
+          default: 100
+        }
+      ]);
+    },
+    afterRender: function(question, el) {
+      if (question.survey.platformName == "vue") return;
+      const ComponentClass = Vue.extend(CustomDate);
+      const card = new ComponentClass({
+        propsData: { question: question }
+      });
+      card.$mount();
+      el.appendChild(card.$el);
+    }
+  };
+  Survey.CustomWidgetCollection.Instance.addCustomWidget(widget, "property");
+}
