+++ web/src/components/survey/question-types/survey-init.ts
@@ -1,21 +1,21 @@
 import { addQuestionTypes, WidgetValueName } from "./question-types";
 import Vue from "vue";
 
-import AddressInfo from "./components/AddressInfo.vue";
-import ContactInfo from "./components/ContactInfo.vue";
-import CustomDate from "./components/CustomDate.vue";
-import HelpText from "./components/HelpText.vue";
-import InfoText from "./components/InfoText.vue";
-import PersonName from "./components/PersonName.vue";
-import YesNo from "./components/YesNo.vue";
-import FormDownloadButton from "./components/FormDownloadButton.vue";
-import ReviewAnswers from "./components/ReviewAnswers.vue";
-import DateMath from "./components/DateMath.vue";
+import AddressInfo from "../components/AddressInfo.vue";
+import ContactInfo from "../components/ContactInfo.vue";
+import CustomDate from "../components/CustomDate.vue";
+import HelpText from "../components/HelpText.vue";
+import InfoText from "../components/InfoText.vue";
+import PersonName from "../components/PersonName.vue";
+import YesNo from "../components/YesNo.vue";
+import FormDownloadButton from "../components/FormDownloadButton.vue";
+import ReviewAnswers from "../components/ReviewAnswers.vue";
+import DateMath from "../components/DateMath.vue";
 
-import EarliestSubmissionDate from "./components/EarliestSubmissionDate.vue";
+import EarliestSubmissionDate from "../components/EarliestSubmissionDate.vue";
 import VRuntimeTemplate from "v-runtime-template";
-import StringViewer from "./components/outer-question/StringViewer.vue";
-import SurveyText from "./components/outer-question/SurveyText.vue";
+import StringViewer from "../components/outer-question/StringViewer.vue";
+import SurveyText from "../components/outer-question/SurveyText.vue";
 
 export function addQuestionTypesVue(Survey: any) {
   Vue.component(WidgetValueName[WidgetValueName.HelpText], HelpText);
