+++ web/src/components/survey/question-types/yes-no-type.ts
@@ -1,0 +1,32 @@
+import Vue from "vue";
+import YesNo from "../components/YesNo.vue";
+import { WidgetValueName } from "./question-types";
+
+export function initYesNo(Survey: any) {
+  const widget = {
+    name: WidgetValueName[WidgetValueName.YesNo],
+    title: "Yes/No",
+    iconName: "icon-radiogroup",
+    isDefaultRender: true,
+    widgetIsLoaded: function() {
+      return true;
+    },
+    isFit: function(question: any) {
+      return question.getType() === "yesno";
+    },
+    activatedByChanged: function(activatedBy: any) {
+      Survey.JsonObject.metaData.addClass("yesno", [], null, "empty");
+    },
+    htmlTemplate: "<div></div>",
+    afterRender: function(question, el) {
+      if (question.survey.platformName == "vue") return;
+      const ComponentClass = Vue.extend(YesNo);
+      const card = new ComponentClass({
+        propsData: { question: question }
+      });
+      card.$mount();
+      el.appendChild(card.$el);
+    }
+  };
+  Survey.CustomWidgetCollection.Instance.addCustomWidget(widget, "type");
+}
